Remove backstage for sm (#122)
* First (hacky) attempt at removing backstage

Without the verify=false on the decode, the decryption wouldn't happen.  This needs to be looked at.
Other then that, the uaa code is just a copy of what is in backstage and seems to work fine once
the decryption is done.

* Move remaining SM functionality off of backstage

* Remove unused config and regen lockfile

* Add limit that used to be added in backstage

* Fix unit tests

* Install crpytography to fix invalidAlgorithm problem

* Fix unused import

* Added uaa tests and removed useless error handling

The keyerror exception block (found in the first try block in view_conversation) wasn't used by anything.
get_conversation doesn't raise keyErrors, the _refine method handles keyErrors by logging out an error message then
allowing the code to resume execution instead of 'raise'ing it and letting it bubble up. and _get_conversation_breadcrumbs
uses .get() so it can't raise a keyError.